# Introduce string encoder to reduce allocations