It's me that added duplication of json encoder and buffer on marshalizing map keys (in #15). I improved the performance of encoding strings by reusing the buffer. The encoder still copies on converting to `string`, but let's start from here. Are you interested in further performance improvements?

```
benchmark              old ns/op     new ns/op     delta
BenchmarkFromat-16     12481         11268         -9.72%

benchmark              old allocs     new allocs     delta
BenchmarkFromat-16     191            173            -9.42%

benchmark              old bytes     new bytes     delta
BenchmarkFromat-16     6103          5078          -16.80%
```