@itchyny 2021-01-12
Hmm, the formatter is expected to work in multiple goroutines?